Change costs to int

+++ ChargeNotificationDbProj/Tables/CustomerGameCharge.sql
@@ -3,7 +3,7 @@
     [Id] INT NOT NULL PRIMARY KEY, 
     [GameName] NVARCHAR(50) NOT NULL, 
     [Description] NVARCHAR(50) NOT NULL, 
-    [TotalCost] FLOAT NOT NULL, 
+    [TotalCost] INT NOT NULL, 
     [ChargeDate] DATE NOT NULL, 
     [CustomerId] INT NOT NULL, 
     CONSTRAINT [FK_CustomerGameCharge_ToTable] FOREIGN KEY ([CustomerId]) REFERENCES [Customer]([Id]),
